removed linux,patch,docker, custom func in source
